fix mobile chart
